// src/types/JobEntry.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type JobEntry = { id: string, insert_date: string, company: string, title: string, link?: string, application_date: string, status: string, insert_status: string, location?: string, note?: string, };